remove old bitstruct, refactor some code
