+++ pymtl3/datatypes/__init__.py
@@ -1,4 +1,4 @@
 from .bits_import import *
 from .bits_import import _bitwidths
-from .BitStruct import BitStruct, mk_bit_struct
+from .bit_struct  import bit_struct, is_bit_struct
 from .helpers import clog2, concat, reduce_and, reduce_or, reduce_xor, sext, zext
